Modified the existing code to group the tasks with respect to group names and assignment of tasks is done only to people of the selected group

+++ Database Scripts/usp_tasks.sql
@@ -22,17 +22,18 @@ CREATE PROCEDURE [dbo].[usp_tasks]
 	@Task_Id NVARCHAR(50)=NULL,
 	@Task_Title NVARCHAR(50)=NULL,
 	@Task_Description NVARCHAR(50)=NULL,
-	@Points NVARCHAR(20) = NULL,
-	@Action NVARCHAR(10) = NULL,
+	@Points NVARCHAR(20) = NULL,	
 	@Task_Type NVARCHAR(50)= NULL,
-	@Is_Task_Recursive NVARCHAR(50)=NULL
+	@Is_Task_Recursive NVARCHAR(50)=NULL,
+	@Group_Name NVARCHAR(50)=NULL,
+	@Action NVARCHAR(20) 
 )
 AS
 BEGIN
 	IF @Action = 'C' -- create/save task details
 	
-		INSERT INTO Tasks ( Task_Title, Task_Description, Points, Task_Type, Is_Task_Recursive)
-				   VALUES ( @Task_Title,@Task_Description,@Points,@Task_Type,@Is_Task_Recursive)
+		INSERT INTO Tasks ( Task_Title, Task_Description, Points, Task_Type, Is_Task_Recursive,Group_Name)
+				   VALUES ( @Task_Title,@Task_Description,@Points,@Task_Type,@Is_Task_Recursive,@Group_Name)
 	ELSE IF @Action = 'U'
 		UPDATE Tasks SET Task_Title=@Task_Title,
 						 Task_Description=@Task_Description,
@@ -42,10 +43,15 @@ BEGIN
 						 WHERE Task_Id=@Task_Id
 	ELSE IF @Action = 'R'
 		SELECT Task_Id, Task_Title, Task_Description, Points,Task_Type,Is_Task_Recursive FROM Tasks
+	ELSE IF @Action='GETTASKSBYGROUP'	
+		SELECT Task_Id, Task_Title, Task_Description, Points,Task_Type,Is_Task_Recursive FROM Tasks
+		WHERE Group_Name=@Group_Name
 		
 	ELSE IF @Action = 'D'
-		DELETE FROM Tasks WHERE Task_Id=@Task_Id
-	
+		BEGIN
+			IF NOT EXISTS (SELECT Task_Id from AssignedTasks where Task_Id=@Task_Id)
+				DELETE FROM Tasks WHERE Task_Id=@Task_Id
+		END
 	ELSE IF @Action= 'UPDATETASK' -- UPDATE THE DATA CORRESPONDING TO A SPECIFIC TASK BASED ON THE TASK_ID	
 		UPDATE Tasks SET Points=@Points,
 						 Task_Title=@Task_Title,
